Revert "Create a parent class Contact to extract commonalities from Phone, Email and Address."
This reverts commit 8532eb800f4d3e4782256ecce22ba2a705e5a45a.
