--- src/seedu/addressbook/data/person/Contact.java
@@ -1,29 +1,0 @@
-package seedu.addressbook.data.person;
-
-public class Contact {
-
-    public String value;
-    protected boolean isPrivate;
-    
-
-    @Override
-    public String toString() {
-        return value;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof Phone // instanceof handles nulls
-                && this.value.equals(((Phone) other).value)); // state check
-    }
-
-    @Override
-    public int hashCode() {
-        return value.hashCode();
-    }
-
-    public boolean isPrivate() {
-        return isPrivate;
-    }
-}
