+++ src/seedu/addressbook/data/person/Email.java
@@ -6,12 +6,15 @@ import seedu.addressbook.data.exception.IllegalValueException;
  * Represents a Person's email in the address book.
  * Guarantees: immutable; is valid as declared in {@link #isValidEmail(String)}
  */
-public class Email extends Contact {
+public class Email {
 
     public static final String EXAMPLE = "valid@e.mail";
     public static final String MESSAGE_EMAIL_CONSTRAINTS =
             "Person emails should be 2 alphanumeric/period strings separated by '@'";
     public static final String EMAIL_VALIDATION_REGEX = "[\\w\\.]+@[\\w\\.]+";
+
+    public final String value;
+    private boolean isPrivate;
 
     /**
      * Validates given email.
@@ -34,5 +37,25 @@ public class Email extends Contact {
         return test.matches(EMAIL_VALIDATION_REGEX);
     }
 
+    @Override
+    public String toString() {
+        return value;
+    }
 
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Email // instanceof handles nulls
+                && this.value.equals(((Email) other).value)); // state check
+    }
+
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
 }
